Move some checks to `SignedBlock::deserialize`
